Rename SQLiteService to DatabaseService to align more /w functionality
In order to not be too opinionated about choice of database technology,
let's rename accordingly
